Split Encode class
